pyproject.toml
[project]
name = "calendar-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk==1.6.1",
    "openlit>=1.33.23",
    "psycopg2-binary>=2.9.10",
    "pytz>=2025.2",
    "uvicorn>=0.34.3",
]
